Refactoring follow-up of [#179381846]
