+++ src/main/java/org/codeforamerica/shiba/output/applicationinputsmappers/OneToManyApplicationInputsMapper.java
@@ -1,9 +1,5 @@
 package org.codeforamerica.shiba.output.applicationinputsmappers;
 
-import static org.codeforamerica.shiba.application.parsers.ApplicationDataParser.Field.HOME_EXPENSES;
-import static org.codeforamerica.shiba.application.parsers.ApplicationDataParser.Field.UNEARNED_INCOME;
-import static org.codeforamerica.shiba.application.parsers.ApplicationDataParser.Field.UNEARNED_INCOME_CCAP;
-import static org.codeforamerica.shiba.application.parsers.ApplicationDataParser.Field.UTILITY_PAYMENTS;
 import static org.codeforamerica.shiba.application.parsers.ApplicationDataParser.getValues;
 import static org.codeforamerica.shiba.output.ApplicationInputType.ENUMERATED_SINGLE_VALUE;
 
@@ -15,56 +11,34 @@ import org.codeforamerica.shiba.output.ApplicationInput;
 import org.codeforamerica.shiba.output.Document;
 import org.codeforamerica.shiba.output.Recipient;
 import org.codeforamerica.shiba.pages.data.PagesData;
-import org.springframework.stereotype.Component;
 
 /**
  * Handle checkbox selections that correspond to individual inputs on the presented application.
  */
-@Component
-public class OneToManyApplicationInputsMapper implements ApplicationInputsMapper {
-
-  private static final List<String> UNEARNED_INCOME_OPTIONS = List.of("SOCIAL_SECURITY", "SSI",
-      "VETERANS_BENEFITS",
-      "UNEMPLOYMENT", "WORKERS_COMPENSATION", "RETIREMENT", "CHILD_OR_SPOUSAL_SUPPORT",
-      "TRIBAL_PAYMENTS");
-
-  private static final List<String> UNEARNED_INCOME_CCAP_OPTIONS = List.of("BENEFITS",
-      "INSURANCE_PAYMENTS", "CONTRACT_FOR_DEED", "TRUST_MONEY", "HEALTH_CARE_REIMBURSEMENT",
-      "INTEREST_DIVIDENDS", "OTHER_SOURCES");
-
-  private static final List<String> HOME_EXPENSES_OPTIONS = List.of("RENT", "MORTGAGE",
-      "HOMEOWNERS_INSURANCE", "REAL_ESTATE_TAXES", "ASSOCIATION_FEES", "ROOM_AND_BOARD");
-
-  private static final List<String> UTILITY_PAYMENTS_OPTIONS = List.of("ELECTRICITY", "PHONE",
-      "GARBAGE_REMOVAL", "COOKING_FUEL");
+public abstract class OneToManyApplicationInputsMapper implements ApplicationInputsMapper {
 
   @Override
-
   public List<ApplicationInput> map(Application application, Document document, Recipient recipient,
       SubworkflowIterationScopeTracker scopeTracker) {
     return map(application.getApplicationData().getPagesData());
   }
 
-  private List<ApplicationInput> map(PagesData pagesData) {
-    List<ApplicationInput> results = new ArrayList<>();
+  protected abstract OneToManyParams getParams();
 
-    addApplicationInputs(pagesData, "unearnedIncome", results, UNEARNED_INCOME,
-        UNEARNED_INCOME_OPTIONS);
-
-    addApplicationInputs(pagesData, "unearnedIncomeCcap", results, UNEARNED_INCOME_CCAP,
-        UNEARNED_INCOME_CCAP_OPTIONS);
-
-    addApplicationInputs(pagesData, "homeExpenses", results, HOME_EXPENSES,
-        HOME_EXPENSES_OPTIONS);
-
-    addApplicationInputs(pagesData, "utilityPayments", results, UTILITY_PAYMENTS,
-        UTILITY_PAYMENTS_OPTIONS);
-
-    return results;
+  protected List<ApplicationInput> map(PagesData pagesData) {
+    var params = getParams();
+    return addApplicationInputs(params, pagesData);
   }
 
-  private void addApplicationInputs(PagesData pagesData, String pageName,
-      List<ApplicationInput> results, Field field, List<String> options) {
+  protected List<ApplicationInput> addApplicationInputs(OneToManyParams params,
+      PagesData pagesData) {
+    return addApplicationInputs(pagesData, params.getPageName(), params.getField(),
+        params.getYesNoOptions());
+  }
+
+  private List<ApplicationInput> addApplicationInputs(PagesData pagesData, String pageName,
+      Field field, List<String> options) {
+    List<ApplicationInput> results = new ArrayList<>();
     if (pagesData.containsKey(pageName)) {
       List<String> selectedValues = getValues(pagesData, field);
       options.stream()
@@ -74,5 +48,7 @@ public class OneToManyApplicationInputsMapper implements ApplicationInputsMapper
               ENUMERATED_SINGLE_VALUE))
           .forEach(results::add);
     }
+    return results;
   }
+
 }
